Add generic method for loading modules with fallbacks
This is based on the most complex use case for this I saw mentioned
(where we are loading JSON modules). I noticed that we additionally
could allow a coderef to define the wanted subroutines, so figured we
may as well add functionality to do that if we wanted to.

Additionally this deduplicates the logic for "safely/conditionally"
loading modules for other dependencies in SSP.